requirements.txt
nfstream>=6.5.0
pandas>=2.0.0
numpy>=1.24.0
click>=8.1.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
requests>=2.31.0
jinja2>=3.1.0
folium>=0.15.0

# opcjonalnie (jak chcesz “prawdziwe” PySigma)
# pysigma>=0.11.0
# pysigma-backend-python>=0.2.0
